Refactor: RacingCarGameController와 index.js를 분리
- RacingCarGameController는 src/controller/RacingCarGameController.js로 위치 이동
